feat: pdf overlay state compression
PDF sync doesn't follow the diff system, as updates are in frequent and long busts. Instead, what is synced are the layers from the particular page of the pdf that is being edited.

I've added a pore-processing step where the data is compressed using the "pako" nom library for a 30-40% estimated reduction in data size. The data is decompressed in the client.

This causes a small overhead in CPU usage, but unless this app is runnin in a potato or a 20+ years old computer, that shouldn't be an issue.

I'm considering whether just sending the specific layer of the modified page, instead of all 3 of them, but I'm not sure if that would increase speed too much.

Next commit should include a system with conflict resolution if two clients are editing the same page at the same time, so changes can be reconciled.