feat: debounce layer updates
Layer updates were causing to send every little change and mouse click separately to the clients, which is okay in a local network but a bit excessive through an internet connection.

I've introduced debouncing, so it only sends updates every 100ms. I think this is a good balance between swiftness and performance.